backend/requirements.txt
# ──────────────────────────────────────────────
# DocBott - Python Dependencies
# ──────────────────────────────────────────────

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
SQLAlchemy==2.0.36

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib==1.7.4

# Environment
python-dotenv==1.0.1

# PDF Processing
PyMuPDF==1.25.1

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.6.0
pytesseract==0.3.13
pdf2image==1.17.0
Pillow>=10.0.0

# Table Extraction
camelot-py[cv]==0.11.0
opencv-python-headless>=4.8.0

# Data Processing
pandas>=2.0.0
nltk>=3.8.0

# Embeddings & Vector Store
sentence-transformers>=3.0.0
chromadb>=0.5.0

# Keyword Search
rank-bm25==0.2.2

# AI / Summarization
transformers>=4.40.0
torch>=2.0.0

# Evaluation Metrics
rouge-score>=0.1.2

# HTTP Client (for Ollama)
requests>=2.31.0

# Optional: Redis Cache
# redis>=5.0.0

# Optional: FAISS
# faiss-cpu>=1.7.0
